perf: implement modular for form service
